refactor(bindings/python): Refactor layout for python bindings
